added convenient audiofile wrappers for playing and caching audio file
	* Makefile.am, esdfile.c, esd.h: added convenient audiofile
	wrappers for playing and caching audio file data.
	* esd.h, esdlib.c: const clean up.
	* esdmgr.c: added name to print info functions.
	* esdplay.c, esdsample.c: updated to use audiofile wrappers.
	NOTE: this means that the AUDIOFILE library is now REQUIRED for
	proper compilation of esound.  configure.in will be updated soon.

+++ esdsample.c
@@ -28,12 +28,14 @@ int main(int argc, char **argv)
     struct stat source_stats;
     char *host = NULL;
     char *name = NULL;
+    char *filename = NULL;
+    int cache_mode = 0; /* cache mode, 1 = data, 0 = file */
     
     for ( arg = 1 ; arg < argc ; arg++)
     {
 	if (!strcmp("-h",argv[arg]))
 	{
-	    printf("usage:\n\t%s [-s server] [-b] [-m] [-r freq] < file\n",
+	    printf("usage:\n\t%s [-s server] [-d [-b] [-m] [-r freq]] < file\n",
 		   argv[0]);
 	    exit(0);
 	}
@@ -43,55 +45,81 @@ int main(int argc, char **argv)
 	    bits = ESD_BITS8;
 	else if ( !strcmp( "-m", argv[ arg ] ) )
 	    channels = ESD_MONO;
+	else if ( !strcmp( "-d", argv[ arg ] ) )
+	    cache_mode = 1;
 	else if ( !strcmp( "-r", argv[ arg ] ) )
 	{
 	    arg++;
 	    rate = atoi( argv[ arg ] );
 	} else {
-	    name = argv[ arg ];
-	    source = fopen( name, "r" );
+	    break;
 	}
     }
 
-    if ( source == NULL ) {
-	fprintf( stderr, "%s, sample file not specified\n", argv[ 0 ] );
-	return -1;
-    }
-    
-    format = bits | channels | mode | func;
-    printf( "opening socket, format = 0x%08x at %d Hz\n", 
-	    format, rate );
-   
-    sock = esd_open_sound( host );
-    if ( sock <= 0 ) 
-	return 1;
-
-    stat( name, &source_stats );
-    sample_id = esd_sample_cache( sock, format, rate, source_stats.st_size, name );
-    printf( "sample id is <%d>\n", sample_id );
+    /* filename is the next option */
+    name = argv[ arg ];
 
     /* if we see any of these, terminate */
     signal( SIGINT, clean_exit );
     signal( SIGKILL, clean_exit );
     signal( SIGPIPE, clean_exit );
 
-    while ( ( length = fread( buf, 1, ESD_BUF_SIZE, source ) ) > 0 )
-    {
-	/* fprintf( stderr, "read %d\n", length ); */
-	if ( ( length = write( sock, buf, length)  ) <= 0 )
+    if ( cache_mode ) {
+	source = fopen( name, "r" );
+
+	if ( source == NULL ) {
+	    fprintf( stderr, "%s, sample file not specified\n", argv[ 0 ] );
+	    return -1;
+	}
+    
+	format = bits | channels | mode | func;
+	printf( "opening socket, format = 0x%08x at %d Hz\n", 
+		format, rate );
+   
+	sock = esd_open_sound( host );
+	if ( sock <= 0 ) 
 	    return 1;
-	else
-	    total += length;
-    }
+	
+	stat( name, &source_stats );
+	sample_id = esd_sample_cache( sock, format, rate, source_stats.st_size, name );
+	printf( "sample id is <%d>\n", sample_id );
+    
+	while ( ( length = fread( buf, 1, ESD_BUF_SIZE, source ) ) > 0 )
+	{
+	    /* fprintf( stderr, "read %d\n", length ); */
+	    if ( ( length = write( sock, buf, length)  ) <= 0 )
+		return 1;
+	    else
+		total += length;
+	}
+	
+	confirm_id = esd_confirm_sample_cache( sock );
 
-    confirm_id = esd_confirm_sample_cache( sock );
-    if ( sample_id != confirm_id ) {
-	printf( "error while caching sample <%d>: confirm returned %d\n",
-		sample_id, confirm_id );
-	exit( 1 );
-    }
-    printf( "sample <%d> uploaded, %d bytes\n", sample_id, total );
+	if ( sample_id != confirm_id ) {
+	    printf( "error while caching sample <%d>: confirm returned %d\n",
+		    sample_id, confirm_id );
+	    exit( 1 );
+	}
 
+	printf( "sample <%d> uploaded, %d bytes\n", sample_id, total );
+    } 
+    else {
+	sock = esd_open_sound( host );
+	if ( sock <= 0 ) 
+	    return 1;
+	
+	sample_id = esd_file_cache( sock, name );
+	printf( "sample id is <%d>\n", sample_id );
+    
+	/* if we see any of these, terminate */
+	if ( sample_id < 0 ) {
+	    printf( "error while caching sample <%d>: confirm value != sample_id, %d\n",
+		    sample_id );
+	    exit( 1 );
+	}
+
+	printf( "sample <%d> uploaded: %s\n", sample_id, name );
+    }
 
     reget_sample_id = esd_sample_getid( sock, name );
